requirements.txt
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
branca==0.6.0
certifi==2023.7.22
charset-normalizer==3.3.0
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
Faker==18.13.0
folium==0.14.0
idna==3.4
install==1.3.5
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
jsonref==1.1.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.0
openrouteservice==2.3.3
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pysmartdatamodels==0.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
rstr==3.2.1
shapely==2.0.1
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.10.1
urllib3==2.0.5
wcwidth==0.2.8
